refactor: update import paths for ActionSpecificData in userActionsDebugTable components
